refactor: experiment with removing static_casts for variant assignment

+++ libtransmission/rpcimpl.cc
@@ -665,7 +665,7 @@ namespace make_torrent_field_helpers
     case TR_KEY_activityDate: return st.activityDate;
     case TR_KEY_addedDate: return st.addedDate;
     case TR_KEY_availability: return make_piece_availability_vec(tor);
-    case TR_KEY_bandwidthPriority: return static_cast<int64_t>(tor.get_priority());
+    case TR_KEY_bandwidthPriority: return tor.get_priority();
     case TR_KEY_comment: return tor.comment();
     case TR_KEY_corruptEver: return st.corruptEver;
     case TR_KEY_creator: return tor.creator();
@@ -673,16 +673,16 @@ namespace make_torrent_field_helpers
     case TR_KEY_desiredAvailable: return st.desiredAvailable;
     case TR_KEY_doneDate: return st.doneDate;
     case TR_KEY_downloadDir: return tor.download_dir().sv();
-    case TR_KEY_downloadLimit: return static_cast<int64_t>(tr_torrentGetSpeedLimit_KBps(&tor, TR_DOWN));
+    case TR_KEY_downloadLimit: return tr_torrentGetSpeedLimit_KBps(&tor, TR_DOWN);
     case TR_KEY_downloadLimited: return tor.uses_speed_limit(TR_DOWN);
     case TR_KEY_downloadedEver: return st.downloadedEver;
     case TR_KEY_editDate: return st.editDate;
-    case TR_KEY_error: return static_cast<int64_t>(st.error);
+    case TR_KEY_error: return st.error;
     case TR_KEY_errorString: return st.errorString;
     case TR_KEY_eta: return st.eta;
     case TR_KEY_etaIdle: return st.etaIdle;
     case TR_KEY_fileStats: return make_file_stats_vec(tor);
-    case TR_KEY_file_count: return static_cast<int64_t>(tor.file_count());
+    case TR_KEY_file_count: return tor.file_count();
     case TR_KEY_files: return make_file_vec(tor);
     case TR_KEY_group: return tor.bandwidth_group().sv();
     case TR_KEY_hashString: return tor.info_hash_string().sv();
@@ -697,23 +697,23 @@ namespace make_torrent_field_helpers
     case TR_KEY_leftUntilDone: return st.leftUntilDone;
     case TR_KEY_magnetLink: return tor.metainfo_.magnet();
     case TR_KEY_manualAnnounceTime: return tr_announcerNextManualAnnounce(&tor);
-    case TR_KEY_maxConnectedPeers: return static_cast<int64_t>(tor.peer_limit());
+    case TR_KEY_maxConnectedPeers: return tor.peer_limit();
     case TR_KEY_metadataPercentComplete: return st.metadataPercentComplete;
     case TR_KEY_name: return tor.name();
-    case TR_KEY_peer_limit: return static_cast<int64_t>(tor.peer_limit());
+    case TR_KEY_peer_limit: return tor.peer_limit();
     case TR_KEY_peers: return make_peer_vec(tor);
-    case TR_KEY_peersConnected: return static_cast<int64_t>(st.peersConnected);
+    case TR_KEY_peersConnected: return st.peersConnected;
     case TR_KEY_peersFrom: return make_peer_counts_map(st);
     case TR_KEY_peersGettingFromUs: return st.peersGettingFromUs;
     case TR_KEY_peersSendingToUs: return st.peersSendingToUs;
     case TR_KEY_percentComplete: return st.percentComplete;
     case TR_KEY_percentDone: return st.percentDone;
-    case TR_KEY_pieceCount: return static_cast<int64_t>(tor.piece_count());
-    case TR_KEY_pieceSize: return static_cast<int64_t>(tor.piece_size());
+    case TR_KEY_pieceCount: return tor.piece_count();
+    case TR_KEY_pieceSize: return tor.piece_size();
     case TR_KEY_pieces: return make_piece_bitfield(tor);
     case TR_KEY_primary_mime_type: return tor.primary_mime_type();
     case TR_KEY_priorities: return make_file_priorities_vec(tor);
-    case TR_KEY_queuePosition: return static_cast<int64_t>(st.queuePosition);
+    case TR_KEY_queuePosition: return st.queuePosition;
     case TR_KEY_rateDownload: return tr_toSpeedBytes(st.pieceDownloadSpeed_KBps);
     case TR_KEY_rateUpload: return tr_toSpeedBytes(st.pieceUploadSpeed_KBps);
     case TR_KEY_recheckProgress: return st.recheckProgress;
@@ -727,13 +727,13 @@ namespace make_torrent_field_helpers
     case TR_KEY_sizeWhenDone: return st.sizeWhenDone;
     case TR_KEY_source: return tor.source();
     case TR_KEY_startDate: return st.startDate;
-    case TR_KEY_status: return static_cast<int64_t>(st.activity);
+    case TR_KEY_status: return st.activity;
     case TR_KEY_torrentFile: return tor.torrent_file();
     case TR_KEY_totalSize: return tor.total_size();
     case TR_KEY_trackerList: return tor.tracker_list();
     case TR_KEY_trackerStats: return make_tracker_stats_vec(tor);
     case TR_KEY_trackers: return make_tracker_vec(tor);
-    case TR_KEY_uploadLimit: return static_cast<int64_t>(tr_torrentGetSpeedLimit_KBps(&tor, TR_UP));
+    case TR_KEY_uploadLimit: return tr_torrentGetSpeedLimit_KBps(&tor, TR_UP);
     case TR_KEY_uploadLimited: return tor.uses_speed_limit(TR_UP);
     case TR_KEY_uploadRatio: return st.ratio;
     case TR_KEY_uploadedEver: return st.uploadedEver;
@@ -1341,7 +1341,7 @@ void onBlocklistFetched(tr_web::FetchResponse const& web_response)
     }
 
     // feed it to the session and give the client a response
-    data->args_out.try_emplace(TR_KEY_blocklist_size, static_cast<int64_t>(tr_blocklistSetContent(session, filename)));
+    data->args_out.try_emplace(TR_KEY_blocklist_size, tr_blocklistSetContent(session, filename));
     tr_sys_path_remove(filename);
     tr_idle_function_done(data, SuccessResult);
 }
@@ -1990,13 +1990,13 @@ char const* sessionStats(
     // clang-format off
     switch (key)
     {
-    case TR_KEY_alt_speed_down: return static_cast<int64_t>(tr_sessionGetAltSpeed_KBps(&session, TR_DOWN));
+    case TR_KEY_alt_speed_down: return tr_sessionGetAltSpeed_KBps(&session, TR_DOWN);
     case TR_KEY_alt_speed_enabled: return tr_sessionUsesAltSpeed(&session);
     case TR_KEY_alt_speed_time_begin: return tr_sessionGetAltSpeedBegin(&session);
-    case TR_KEY_alt_speed_time_day: return static_cast<int64_t>(tr_sessionGetAltSpeedDay(&session));
+    case TR_KEY_alt_speed_time_day: return tr_sessionGetAltSpeedDay(&session);
     case TR_KEY_alt_speed_time_enabled: return tr_sessionUsesAltSpeedTime(&session);
     case TR_KEY_alt_speed_time_end: return tr_sessionGetAltSpeedEnd(&session);
-    case TR_KEY_alt_speed_up: return static_cast<int64_t>(tr_sessionGetAltSpeed_KBps(&session, TR_UP));
+    case TR_KEY_alt_speed_up: return tr_sessionGetAltSpeed_KBps(&session, TR_UP);
     case TR_KEY_anti_brute_force_enabled: return tr_sessionGetAntiBruteForceEnabled(&session);
     case TR_KEY_anti_brute_force_threshold: return tr_sessionGetAntiBruteForceThreshold(&session);
     case TR_KEY_blocklist_enabled: return session.blocklist_enabled();
@@ -2039,9 +2039,9 @@ char const* sessionStats(
     case TR_KEY_seed_queue_enabled: return session.queueEnabled(TR_UP);
     case TR_KEY_seed_queue_size: return session.queueSize(TR_UP);
     case TR_KEY_session_id: return session.sessionId();
-    case TR_KEY_speed_limit_down: return static_cast<int64_t>(tr_sessionGetSpeedLimit_KBps(&session, TR_DOWN));
+    case TR_KEY_speed_limit_down: return tr_sessionGetSpeedLimit_KBps(&session, TR_DOWN);
     case TR_KEY_speed_limit_down_enabled: return session.isSpeedLimited(TR_DOWN);
-    case TR_KEY_speed_limit_up: return static_cast<int64_t>(tr_sessionGetSpeedLimit_KBps(&session, TR_UP));
+    case TR_KEY_speed_limit_up: return tr_sessionGetSpeedLimit_KBps(&session, TR_UP);
     case TR_KEY_speed_limit_up_enabled: return session.isSpeedLimited(TR_UP);
     case TR_KEY_start_added_torrents: return !session.shouldPauseAddedTorrents();
     case TR_KEY_tcp_enabled: return session.allowsTCP();
